+++ src/Emulator/Main/Utilities/GDB/PacketData.cs
@@ -47,7 +47,7 @@ namespace Antmicro.Renode.Utilities.GDB
                                                 !address.HasValue ? string.Empty : string.Format("{0:X2}", address)));
         }
 
-        public static PacketData StopReply(BreakpointType reason, uint cpuId, ulong? address)
+        public static PacketData StopReply(BreakpointType reason, ulong cpuId, ulong? address)
         {
             return new PacketData(string.Format("T05{0}:{1};thread:{2:X2};", reason.GetStopReason(),
                                                 !address.HasValue ? string.Empty : string.Format("{0:X2}", address), cpuId));
